Переработка приложения на SingleActivity: - MediaActivity переписана на Fragment - В MainActivity добавлен NavController, подключен BottomNavigationView.
