+++ app/src/main/java/com/example/playlistmaker/ui/media/MediaFragment.kt
@@ -1,0 +1,60 @@
+package com.example.playlistmaker.ui.media
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.FragmentMediaBinding
+import com.example.playlistmaker.ui.media.vp_adapter.AdapterVP
+import com.google.android.material.tabs.TabLayoutMediator
+
+class MediaFragment : Fragment() {
+    private var _binding: FragmentMediaBinding? = null
+    private val binding: FragmentMediaBinding
+        get() = _binding!!
+
+
+    private val tabMediator by lazy {
+        TabLayoutMediator(binding.mediaTl, binding.mediaVp) { tab,position ->
+            tab.text =
+                if (position == 0) getString(R.string.favourite_tracks)
+                else getString(R.string.playlists)
+        }
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentMediaBinding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.mediaVp.adapter = AdapterVP(parentFragmentManager, lifecycle)
+        tabMediator.attach()
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+
+        tabMediator.detach()
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+    }
+
+}
