+++ app/src/main/res/layout/fragment_media.xml
@@ -4,11 +4,11 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.media.MediaActivity">
+    tools:context=".ui.media.MediaFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
-        style="@style/ToolbarAppearance.WithIcon"
+        style="@style/ToolbarAppearance"
         app:title="@string/media"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
