+++ app/src/main/res/menu/nav_menu.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <item android:id="@+id/mediaFragment"
-        android:title="@string/media"
-        android:icon="@drawable/media" />
-
     <item android:id="@+id/searchFragment"
         android:title="@string/search"
         android:icon="@drawable/search" />
+
+    <item android:id="@+id/mediaFragment"
+        android:title="@string/media"
+        android:icon="@drawable/media" />
 
     <item android:id="@+id/settingsFragment"
         android:title="@string/settings"
